+++ src/main/java/video/api/java/sdk/domain/chapter/ChapterInput.java
@@ -1,0 +1,9 @@
+package video.api.java.sdk.domain.chapter;
+
+public class ChapterInput {
+    public String language;
+
+    public ChapterInput(String language) {
+        this.language = language;
+    }
+}